Simplified hierarchy of Google Maps layer view classes
