--- src/geo/gmaps/gmaps-cartodb-layer-view.js
@@ -1,96 +1,0 @@
-var _ = require('underscore');
-var CartoDBLayerGMaps = require('./cartodb-layer-gmaps');
-var GMapsLayerView = require('./gmaps-layer-view');
-
-/**
- * gmaps cartodb layer view
- */
-var GMapsCartoDBLayerView = function(layerModel, gmapsMap) {
-  var self = this;
-
-  _.bindAll(this, 'featureOut', 'featureOver', 'featureClick');
-
-  var opts = _.clone(layerModel.attributes);
-
-  opts.map =  gmapsMap;
-
-  var // preserve the user's callbacks
-  _featureOver  = opts.featureOver,
-  _featureOut   = opts.featureOut,
-  _featureClick = opts.featureClick;
-
-  opts.featureOver  = function() {
-    _featureOver  && _featureOver.apply(this, arguments);
-    self.featureOver  && self.featureOver.apply(this, arguments);
-  };
-
-  opts.featureOut  = function() {
-    _featureOut  && _featureOut.apply(this, arguments);
-    self.featureOut  && self.featureOut.apply(this, arguments);
-  };
-
-  opts.featureClick  = function() {
-    _featureClick  && _featureClick.apply(this, arguments);
-    self.featureClick  && self.featureClick.apply(opts, arguments);
-  };
-
-  CartoDBLayerGMaps.call(this, opts);
-  GMapsLayerView.call(this, layerModel, this, gmapsMap);
-};
-
-_.extend(
-  GMapsCartoDBLayerView.prototype,
-  CartoDBLayerGMaps.prototype,
-  GMapsLayerView.prototype,
-  {
-
-  _update: function() {
-    this.setOptions(this.model.attributes);
-  },
-
-  reload: function() {
-    this.model.invalidate();
-  },
-
-  remove: function() {
-    GMapsLayerView.prototype.remove.call(this);
-    this.clear();
-  },
-
-  featureOver: function(e, latlon, pixelPos, data) {
-    // dont pass gmaps LatLng
-    this.trigger('featureOver', e, [latlon.lat(), latlon.lng()], pixelPos, data, 0);
-  },
-
-  featureOut: function(e) {
-    this.trigger('featureOut', e);
-  },
-
-  featureClick: function(e, latlon, pixelPos, data, layer) {
-    // dont pass leaflet lat/lon
-    this.trigger('featureClick', e, [latlon.lat(), latlon.lng()], pixelPos, data, 0);
-  },
-
-  error: function(e) {
-    if(this.model) {
-      //trigger the error form _checkTiles in the model
-      this.model.trigger('error', e?e.error:'unknown error');
-      this.model.trigger('tileError', e?e.error:'unknown error');
-    }
-  },
-
-  tilesOk: function(e) {
-    this.model.trigger('tileOk');
-  },
-
-  loading: function() {
-    this.trigger("loading");
-  },
-
-  finishLoading: function() {
-    this.trigger("load");
-  }
-
-});
-
-module.exports = GMapsCartoDBLayerView;
