--- src/geo/gmaps/gmaps-cartodb-named-map-view.js
@@ -1,6 +1,0 @@
-var LayerGroupView = require('./gmaps-layer-group-view-fn')
-var CartoDBNamedMapGmaps = require('./cartodb-named-map-gmaps');
-
-var GMapsCartoDBNamedMapView = LayerGroupView(CartoDBNamedMapGmaps);
-
-module.exports = GMapsCartoDBNamedMapView;
