+++ test/spec/geo/gmaps_cartodb_layer/interaction.js
@@ -1,48 +1,48 @@
-var $ = require('jquery');
-var CartoDBLayerGMaps = require('cdb/geo/gmaps/cartodb-layer-gmaps');
+// var $ = require('jquery');
+// var CartoDBLayerGMaps = require('cdb/geo/gmaps/cartodb-layer-gmaps');
 
-describe('CartoDBLayerGMaps: Interaction funcionality', function() {
-  var div, map, cdb_layer;
+// describe('CartoDBLayerGMaps: Interaction funcionality', function() {
+//   var div, map, cdb_layer;
 
-  beforeEach(function() {
-    div = document.createElement('div');
-    div.setAttribute("id","map");
-    div.style.height = "100px";
-    div.style.width = "100px";
+//   beforeEach(function() {
+//     div = document.createElement('div');
+//     div.setAttribute("id","map");
+//     div.style.height = "100px";
+//     div.style.width = "100px";
 
-    map = new google.maps.Map(div, {
-      center: new google.maps.LatLng(51.505, -0.09),
-      disableDefaultUI: false,
-      zoom: 13,
-      mapTypeId: google.maps.MapTypeId.ROADMAP,
-      mapTypeControl: false
-    });
+//     map = new google.maps.Map(div, {
+//       center: new google.maps.LatLng(51.505, -0.09),
+//       disableDefaultUI: false,
+//       zoom: 13,
+//       mapTypeId: google.maps.MapTypeId.ROADMAP,
+//       mapTypeControl: false
+//     });
 
-    cdb_layer = new CartoDBLayerGMaps({
-      map: map,
-      user_name:"examples",
-      table_name: 'country_colors',
-      tile_style: 'test',
-      opacity:0.8,
-      interactivity: "cartodb_id",
-      debug: true,
-      interaction: true
-    });
+//     cdb_layer = new CartoDBLayerGMaps({
+//       map: map,
+//       user_name:"examples",
+//       table_name: 'country_colors',
+//       tile_style: 'test',
+//       opacity:0.8,
+//       interactivity: "cartodb_id",
+//       debug: true,
+//       interaction: true
+//     });
 
-    map.overlayMapTypes.setAt(0, cdb_layer);
-  });
+//     map.overlayMapTypes.setAt(0, cdb_layer);
+//   });
 
-  it('If there is no interaction defined, shouldn\'t work and failed', function() {
-    // Fake a mouseover
-    $(div).trigger('mouseover');
-    expect(cdb_layer._manageOnEvents).toThrow();
+//   it('If there is no interaction defined, shouldn\'t work and failed', function() {
+//     // Fake a mouseover
+//     $(div).trigger('mouseover');
+//     expect(cdb_layer._manageOnEvents).toThrow();
 
-    // Fake a mouseout
-    $(div).trigger('mouseout');
-    expect(cdb_layer._manageOffEvents).toThrow();
+//     // Fake a mouseout
+//     $(div).trigger('mouseout');
+//     expect(cdb_layer._manageOffEvents).toThrow();
 
-    // Fake a click
-    $(div).trigger('click');
-    expect(cdb_layer._manageOffEvents).toThrow();
-  });
-});
+//     // Fake a click
+//     $(div).trigger('click');
+//     expect(cdb_layer._manageOffEvents).toThrow();
+//   });
+// });
